Actualizacion mas grande del programa, algunos cambios de columnas y poner variables que necesitamos
Hace falta arreglar tablas SQL para que el programa corra algunas tablas, pero la estructura se cambio

+++ ProyectoPrograIII/src/java/modelos/Pedido.java
@@ -5,7 +5,7 @@ public class Pedido {
     private int idMesa;
     private int idUsuario;
     private int idCliente;
-    private float Total;
+    private double Total;
     private String Observacion;
     private int Estado;
 
@@ -54,7 +54,7 @@ public class Pedido {
         this.idCliente = idCliente;
     }
 
-    public float getTotal() {
+    public double getTotal() {
         return Total;
     }
